SEQUOIADBMAINSTREAM-3650 消除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@35997 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/test/sql/sqlTest2.cpp
@@ -63,7 +63,7 @@ BOOLEAN readLine ( const char* prompt, char* p, int length )
    ret = getNextCommand( prompt, &readstr ) ;
    if ( readstr )
    {
-      if ( ossStrlen( readstr ) >= length )
+      if ( ossStrlen( readstr ) >= (unsigned int)length )
       {
          cout << "input is to long" << std::endl ;
          ret = FALSE ;
